requirements.txt
streamlit
python-docx
docx2pdf
Pillow
img2pdf
PyPDF2
python-dateutil
pywin32
pythoncom
locale
google-api-python-client
google-auth-httplib2
google-auth-oauthlib 